+++ examples/Playground.hs
@@ -7,9 +7,9 @@ import Universum
 import Data.Monoid ((<>))
 import Data.Yaml.Pretty (defConfig, encodePretty)
 
-import System.Wlog (CanLog, buildAndSetupYamlLogging, logDebug, logError, logInfo, logNotice,
-                    logWarning, modifyLoggerName, parseLoggerConfig, prefixB, productionB,
-                    removeAllHandlers, usingLoggerName)
+import System.Wlog (CanLog, defaultConfig, launchFromFile, launchSimpleLogging, logDebug, logError,
+                    logInfo, logNotice, logWarning, modifyLoggerName, parseLoggerConfig,
+                    productionB, usingLoggerName)
 
 testLoggerConfigPath :: FilePath
 testLoggerConfigPath = "logger-config-example.yaml"
@@ -17,7 +17,7 @@ testLoggerConfigPath = "logger-config-example.yaml"
 testToJsonConfigOutput :: MonadIO m => m ()
 testToJsonConfigOutput = do
     cfg             <- parseLoggerConfig testLoggerConfigPath
-    let builtConfig  = cfg <> productionB <> prefixB "logs"
+    let builtConfig  = cfg <> productionB
     putStrLn $ encodePretty defConfig builtConfig
 
 testLogging :: (CanLog m) => m ()
@@ -33,7 +33,7 @@ testLogging = usingLoggerName "node" $ do
 
     logError   "BARDAQ"
 
-showSomeLog :: IO ()
+showSomeLog :: (CanLog m, MonadIO m) => m ()
 showSomeLog = do
     putText "Other log:"
     usingLoggerName "naked" $ do
@@ -43,8 +43,11 @@ showSomeLog = do
 main :: IO ()
 main = do
     testToJsonConfigOutput
-    let config = (productionB <> prefixB "logs")
-    bracket_
-        (buildAndSetupYamlLogging config testLoggerConfigPath)
-        removeAllHandlers
-        (testLogging >> showSomeLog)
+    let runPlayLog = testLogging >> showSomeLog
+
+    putStrLn $ encodePretty defConfig $ defaultConfig "example"
+    putText "Default configurations.."
+    launchSimpleLogging "node" runPlayLog
+
+    putText "\nFrom file configurations.."
+    launchFromFile testLoggerConfigPath "node" runPlayLog