+++ CHANGES.md
@@ -2,7 +2,7 @@
 =====
 
 * [#61](https://github.com/serokell/log-warper/issues/61):
-  Add `runLoggingWithFile` and `runDefaultLogging` functions.
+  Add `launchFromFile`, `defaultConfig` and `launchSimpleLogging` functions.
 
 1.7.2
 =====